refactor(#186): SIP-0097 slice 1 — hoist pure helpers out of DispatchedFlowExecutor
Per SIP-0097 §6.5 / §8 slice 1 (pure moves, no state threading, no
behavior change):

- src/squadops/cycles/run_report_builder.py: run-report markdown assembly
  (build_run_report + the three line builders); executor's
  _generate_run_report keeps fetch + vault write, delegates formatting
- src/squadops/cycles/failure_evidence.py: build_failure_evidence,
  compose_failure_trigger (verbatim)
- src/squadops/cycles/agent_config.py: ResolvedAgentConfig frozen
  dataclass + resolve_agent_config + build_agent_resolver; consolidates
  the task_plan._resolve_agent_config mirror (#110/#151) — positional
  tuple destructuring retired at all 3 call sites
- adapters/cycles/execution_errors.py: the 4 internal control-flow error
  classes, shared by both executors (in_process duplication removed);
  old import path still works (re-exported via the executor's imports)

Tests: TestBuildFailureEvidence/TestComposeFailureTrigger moved verbatim
to tests/unit/cycles/test_failure_evidence.py (call syntax only);
resolve_agent_config fallback test moved from test_correction_protocol
to new test_agent_config.py + happy-path/None-profile/resolver cases.

Executor: 3,358 -> 3,122 lines. Regression: 4,660 passed, exit 0.

Refs #186 (SIP-0097 slice 1 of 6), #110, #151
